services/notification_service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
psycopg2-binary
pydantic[email]
fastapi-utils
typing-inspect
httpx
pydantic-settings